+++ test/unit/fuzz-tools.test.js
@@ -131,6 +131,7 @@ import * as uroV130 from '../../lib/uro-v130.js';
 import * as uroV131 from '../../lib/uro-v131.js';
 import * as hemeV132 from '../../lib/heme-v132.js';
 import * as warfarinV133 from '../../lib/warfarin-v133.js';
+import * as oncV134 from '../../lib/onc-v134.js';
 import * as emsV149 from '../../lib/ems-v149.js';
 
 const MODULES = {
@@ -207,6 +208,7 @@ const MODULES = {
   'uro-v131.js': uroV131,
   'heme-v132.js': hemeV132,
   'warfarin-v133.js': warfarinV133,
+  'onc-v134.js': oncV134,
   'ems-v149.js': emsV149,
 };